4.회사보충/4. 서버/llm/Demo/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
async-timeout==5.0.1
attrs==24.2.0
blinker==1.8.2
bm25s==0.2.4
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
compressed-tensors==0.6.0
datasets==3.1.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
einops==0.8.0
exceptiongroup==1.2.2
faiss-cpu==1.8.0.post1
fastapi==0.112.4
filelock==3.16.1
Flask==3.0.3
frozenlist==1.5.0
fsspec==2024.9.0
gguf==0.10.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
interegular==0.3.3
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwipiepy==0.20.2
kiwipiepy_model==0.20.0
lark==1.2.2
llvmlite==0.41.1
lm-format-enforcer==0.10.6
MarkupSafe==2.1.5
mistral_common==1.5.1
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.18.6
multidict==6.1.0
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.1
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
openai==1.55.0
opencv-python-headless==4.10.0.84
outlines==0.0.46
packaging==24.2
pandas==2.0.3
partial-json-parser==0.2.1.1.post4
pillow==10.4.0
pip==24.2
pkgutil_resolve_name==1.3.10
prometheus_client==0.21.0
prometheus-fastapi-instrumentator==7.0.0
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
py-cpuinfo==9.0.0
pyairports==2.1.1
pyarrow==17.0.0
pycountry==24.6.1
pydantic==2.10.1
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
ray==2.10.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rpds-py==0.20.1
safetensors==0.4.5
scipy==1.10.1
sentencepiece==0.2.0
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
sympy==1.13.3
tiktoken==0.7.0
tokenizers==0.20.3
torch==2.4.0
torchvision==0.19.0
tqdm==4.67.0
transformers==4.46.3
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
vllm==0.6.3.post1
watchfiles==0.24.0
websockets==13.1
Werkzeug==3.0.6
wheel==0.44.0
xformers==0.0.27.post2
xxhash==3.5.0
yarl==1.15.2
zipp==3.20.2
